refactor: outsource vertex definitions
