+++ src/dataflow/graph/index.ts
@@ -1,2 +1,4 @@
 export * from './graph'
 export * from './edge'
+export * from './vertex'
+export * from './equal'
